Optimize Pokemon D and Z scripts

+++ catscripts/Entertainment/Cartoons/My_Little_Pony:_Friendship_Is_Magic/MLP.sh
@@ -1,15 +1,31 @@
 #!/bin/bash
 #Script fot the My Little Pony: Friendship is Magic cartoon
 
-egrep -i 'Friendship is Magic|MLP\:FIM|MLP FIM|brony|Rainbow Dash|Twilight Sparkle|Pinkie Pie|Princess Celestia|Princess Luna|Princess Cadance|Princess Cadence|Queen Chrysalis|Fluttershy' newpages.txt >> MLPFIM.txt
+KEYWORDS_MLP="Friendship(| )is(| )Magic|MLP\:(| )FIM|MLP(| )FIM|brony|Rainbow(| )Dash|Twilight(| )Sparkle|Pinkie(| )Pie|Princess(| )Celestia|Princess(| )Luna|Princess(| )Cad(a|e)nce|Queen(| )Chrysalis|Fluttershy"
 
-MLP=`stat --print=%s MLPFIM.txt`
-
-if [ $MLP -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="MLPFIM.txt"
-  export CATNAME="My Little Pony: Friendship is Magic"
-  $CATEGORIZE
-fi
 
-rm MLPFIM.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting My Little Pony: Friendship Is Magic\n"
+  fi
+
+  MLP="$(egrep -i "$KEYWORDS_MLP" newpages.txt)"
+
+  if [ "$MLP" != "" ];
+  then
+    printf "%s" "$MLP" > MLPFIM.txt
+    export CATFILE="MLPFIM.txt"
+    export CATNAME="My Little Pony: Friendship is Magic"
+    $CATEGORIZE
+    rm MLPFIM.txt
+    unset MLP
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing My Little Pony: Friendship Is Magic\n"
+  fi
+
+fi
