+++ catscripts/Government/Countries/Portugal/Portugal.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i 'Portugal' newpages.txt >> Portugal.txt
+KEYWORDS_PORTUGAL="Portugal"
 
-PORTUGAL=`stat --print=%s Portugal.txt`
-
-if [ $PORTUGAL -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Portugal.txt"
-  export CATNAME="Portugal"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Portugal\n"
+  fi
 
-rm Portugal.txt
+  PORTUGAL="$(egrep -i "$KEYWORDS_PORTUGAL" newpages.txt)"
+
+  if [ "$PORTUGAL" != "" ];
+  then
+    printf "%s" "$PORTUGAL" > Portugal.txt
+    export CATFILE="Portugal.txt"
+    export CATNAME="Portugal"
+    $CATEGORIZE
+    rm Portugal.txt
+    unset PORTUGAL
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Portugal\n"
+  fi
+
+fi
