+++ catscripts/Lifestyle/Health/Diseases/Cancer/Leukemia/Leukemia.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i 'Leukemia' newpages.txt > Leukemia.txt
+KEYWORDS_LEUKEMIA="Leukemia"
 
-LEUKEMIA=`stat --print=%s Leukemia.txt`
-
-if [ $LEUKEMIA -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Leukemia.txt"
-  export CATNAME="Leukemia"
-  $CATEGORIZE
-fi
+  
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Leukemia\n"
+  fi
 
-rm Leukemia.txt
+  LEUKEMIA="$(egrep -i "$KEYWORDS_LEUKEMIA" newpages.txt)"
+
+  if [ "$LEUKEMIA" != "" ];
+  then
+    printf "%s" "$LEUKEMIA" > Leukemia.txt
+    export CATFILE="Leukemia.txt"
+    export CATNAME="Leukemia"
+    $CATEGORIZE
+    rm Leukemia.txt
+    unset LEUKEMIA
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Leukemia\n"
+  fi
+
+fi
